SNOW-1229108 Refactor config validation
