+++ src/test/java/com/snowflake/kafka/connector/ConnectorConfigValidatorTest.java
@@ -37,7 +37,10 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class ConnectorConfigTest {
+public class ConnectorConfigValidatorTest {
+
+  private final IConnectorConfigValidator connectorConfigValidator = new ConnectorConfigValidator();
+
   // subset of valid community converters
   public static final List<Converter> COMMUNITY_CONVERTER_SUBSET =
       Arrays.asList(
@@ -59,7 +62,7 @@ public class ConnectorConfigTest {
   @Test
   public void testConfig() {
     Map<String, String> config = getConfig();
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -69,14 +72,14 @@ public class ConnectorConfigTest {
 
     Utils.convertAppName(config);
 
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testEmptyFlushTime() {
     Map<String, String> config = getConfig();
     config.remove(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
   }
@@ -87,7 +90,7 @@ public class ConnectorConfigTest {
     config.put(
         SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC,
         (SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC_MIN - 1) + "");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
   }
@@ -96,7 +99,7 @@ public class ConnectorConfigTest {
   public void testFlushTimeNotNumber() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC, "fdas");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
   }
@@ -115,7 +118,7 @@ public class ConnectorConfigTest {
   public void shouldThrowExForEmptyProperty(String prop) {
     Map<String, String> config = getConfig();
     config.remove(prop);
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(prop);
   }
@@ -125,14 +128,14 @@ public class ConnectorConfigTest {
     Map<String, String> config = getConfig();
     config.put(JVM_PROXY_HOST, "127.0.0.1");
     config.put(JVM_PROXY_PORT, "3128");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testEmptyPort() {
     Map<String, String> config = getConfig();
     config.put(JVM_PROXY_HOST, "127.0.0.1");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(JVM_PROXY_HOST);
   }
@@ -141,7 +144,7 @@ public class ConnectorConfigTest {
   public void testEmptyHost() {
     Map<String, String> config = getConfig();
     config.put(JVM_PROXY_PORT, "3128");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(JVM_PROXY_PORT);
   }
@@ -177,7 +180,7 @@ public class ConnectorConfigTest {
   public void testIllegalTopicMap() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.TOPICS_TABLES_MAP, "$@#$#@%^$12312");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.TOPICS_TABLES_MAP);
   }
@@ -186,7 +189,7 @@ public class ConnectorConfigTest {
   public void testIllegalTableName() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.TOPICS_TABLES_MAP, "topic1:!@#@!#!@");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .matches(
             ex ->
@@ -199,7 +202,7 @@ public class ConnectorConfigTest {
   public void testDuplicatedTopic() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.TOPICS_TABLES_MAP, "topic1:table1,topic1:table2");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .matches(
             ex ->
@@ -212,7 +215,7 @@ public class ConnectorConfigTest {
   public void testDuplicatedTableName() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.TOPICS_TABLES_MAP, "topic1:table1,topic2:table1");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -220,14 +223,14 @@ public class ConnectorConfigTest {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.TOPICS, "!@#,$%^,test");
     config.put(SnowflakeSinkConnectorConfig.TOPICS_TABLES_MAP, "!@#:table1,$%^:table2");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testBufferSizeRange() {
     Map<String, String> config = getConfig();
     config.put(BUFFER_SIZE_BYTES, SnowflakeSinkConnectorConfig.BUFFER_SIZE_BYTES_MIN - 1 + "");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_SIZE_BYTES);
   }
@@ -236,7 +239,7 @@ public class ConnectorConfigTest {
   public void testBufferSizeValue() {
     Map<String, String> config = getConfig();
     config.put(BUFFER_SIZE_BYTES, "afdsa");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_SIZE_BYTES);
   }
@@ -245,7 +248,7 @@ public class ConnectorConfigTest {
   public void testEmptyBufferCountNegative() {
     Map<String, String> config = getConfig();
     config.put(BUFFER_COUNT_RECORDS, "-1");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_COUNT_RECORDS);
   }
@@ -254,7 +257,7 @@ public class ConnectorConfigTest {
   public void testBufferCountValue() {
     Map<String, String> config = getConfig();
     config.put(BUFFER_COUNT_RECORDS, "adssadsa");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_COUNT_RECORDS);
   }
@@ -263,34 +266,34 @@ public class ConnectorConfigTest {
   public void testKafkaProviderConfigValue_valid_null() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG, null);
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testKafkaProviderConfigValue_valid_empty() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG, "");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testKafkaProviderConfigValue_valid_provider() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG, "self_hosted");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG, "CONFLUENT");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG, "UNKNOWN");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testKafkaProviderConfigValue_invalid_value() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG, "Something_which_is_not_supported");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.PROVIDER_CONFIG);
   }
@@ -299,17 +302,17 @@ public class ConnectorConfigTest {
   public void testBehaviorOnNullValuesConfig_valid_value() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.BEHAVIOR_ON_NULL_VALUES_CONFIG, "IGNORE");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(SnowflakeSinkConnectorConfig.BEHAVIOR_ON_NULL_VALUES_CONFIG, "DEFAULT");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testBehaviorOnNullValuesConfig_invalid_value() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.BEHAVIOR_ON_NULL_VALUES_CONFIG, "INVALID");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BEHAVIOR_ON_NULL_VALUES_CONFIG);
   }
@@ -318,17 +321,17 @@ public class ConnectorConfigTest {
   public void testJMX_valid_value() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.JMX_OPT, "true");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(SnowflakeSinkConnectorConfig.JMX_OPT, "False");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testJMX_invalid_value() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.JMX_OPT, "INVALID");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.JMX_OPT);
   }
@@ -339,7 +342,7 @@ public class ConnectorConfigTest {
     config.put(
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE.toString());
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -350,7 +353,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -361,7 +364,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(Utils.SF_ROLE);
   }
@@ -370,7 +373,7 @@ public class ConnectorConfigTest {
   public void testIngestionTypeConfig_invalid_value() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT, "INVALID_VALUE");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT);
   }
@@ -421,14 +424,14 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(
         ERRORS_TOLERANCE_CONFIG, SnowflakeSinkConnectorConfig.ErrorTolerance.NONE.toString());
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(ERRORS_TOLERANCE_CONFIG, "all");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -439,7 +442,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT);
   }
@@ -452,13 +455,13 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(ERRORS_LOG_ENABLE_CONFIG, "FALSE");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
 
     config.put(ERRORS_LOG_ENABLE_CONFIG, "TRUE");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -469,7 +472,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT,
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.ERRORS_LOG_ENABLE_CONFIG);
   }
@@ -483,7 +486,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.remove(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
   }
@@ -498,7 +501,7 @@ public class ConnectorConfigTest {
     config.put(
         SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC,
         (StreamingUtils.STREAMING_BUFFER_FLUSH_TIME_MINIMUM_SEC - 1) + "");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
   }
@@ -511,7 +514,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.put(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC, "fdas");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_FLUSH_TIME_SEC);
   }
@@ -524,7 +527,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.remove(BUFFER_SIZE_BYTES);
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.BUFFER_SIZE_BYTES);
   }
@@ -537,7 +540,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.remove(BUFFER_COUNT_RECORDS);
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_COUNT_RECORDS);
   }
@@ -550,7 +553,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.put(BUFFER_COUNT_RECORDS, "-1");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_COUNT_RECORDS);
   }
@@ -563,7 +566,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.put(BUFFER_COUNT_RECORDS, "adssadsa");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(BUFFER_COUNT_RECORDS);
   }
@@ -581,7 +584,7 @@ public class ConnectorConfigTest {
           config.put(
               SnowflakeSinkConnectorConfig.KEY_CONVERTER_CONFIG_FIELD,
               converter.getClass().toString());
-          Utils.validateConfig(config);
+          connectorConfigValidator.validateConfig(config);
         });
 
     COMMUNITY_CONVERTER_SUBSET.forEach(
@@ -589,7 +592,7 @@ public class ConnectorConfigTest {
           config.put(
               SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD,
               converter.getClass().toString());
-          Utils.validateConfig(config);
+          connectorConfigValidator.validateConfig(config);
         });
   }
 
@@ -608,7 +611,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD,
         "org.apache.kafka.connect.storage.StringConverter");
 
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.KEY_CONVERTER_CONFIG_FIELD);
   }
@@ -628,7 +631,7 @@ public class ConnectorConfigTest {
     config.put(
         SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD, converter.getClass().getName());
 
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD);
   }
@@ -649,7 +652,7 @@ public class ConnectorConfigTest {
     config.put(prop, value);
 
     // WHEN/THEN
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @ParameterizedTest
@@ -671,7 +674,7 @@ public class ConnectorConfigTest {
     config.put(prop, value);
 
     // WHEN/THEN
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(prop);
   }
@@ -692,7 +695,7 @@ public class ConnectorConfigTest {
     config.put(prop, value);
 
     // WHEN/THEN
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(prop);
   }
@@ -705,7 +708,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE.toString());
     config.put(SnowflakeSinkConnectorConfig.ENABLE_SCHEMATIZATION_CONFIG, "true");
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
   }
@@ -718,7 +721,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE_STREAMING.toString());
     config.put(SnowflakeSinkConnectorConfig.ENABLE_SCHEMATIZATION_CONFIG, "true");
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -732,7 +735,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD,
         "org.apache.kafka.connect.storage.StringConverter");
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining("org.apache.kafka.connect.storage.StringConverter");
   }
@@ -748,7 +751,7 @@ public class ConnectorConfigTest {
         SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD,
         "org.apache.kafka.connect.storage.StringConverter");
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -760,7 +763,7 @@ public class ConnectorConfigTest {
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.put(SnowflakeSinkConnectorConfig.ENABLE_STREAMING_CLIENT_OPTIMIZATION_CONFIG, "true");
 
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -771,7 +774,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE.toString());
     config.put(SnowflakeSinkConnectorConfig.ENABLE_STREAMING_CLIENT_OPTIMIZATION_CONFIG, "true");
 
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(
             SnowflakeSinkConnectorConfig.ENABLE_STREAMING_CLIENT_OPTIMIZATION_CONFIG);
@@ -786,7 +789,7 @@ public class ConnectorConfigTest {
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.put(SnowflakeSinkConnectorConfig.ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG, "true");
 
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
@@ -797,7 +800,7 @@ public class ConnectorConfigTest {
         IngestionMethodConfig.SNOWPIPE.toString());
     config.put(SnowflakeSinkConnectorConfig.ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG, "true");
 
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(
             SnowflakeSinkConnectorConfig.ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG);
@@ -813,7 +816,7 @@ public class ConnectorConfigTest {
     config.put(Utils.SF_ROLE, "ACCOUNTADMIN");
     config.put(
         SnowflakeSinkConnectorConfig.ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG, "INVALID");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(
             SnowflakeSinkConnectorConfig.ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG);
@@ -828,7 +831,7 @@ public class ConnectorConfigTest {
     config.put(
         SnowflakeSinkConnectorConfig.SNOWPIPE_STREAMING_CLIENT_PROVIDER_OVERRIDE_MAP, "a:b,c:d");
 
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(
             SnowflakeSinkConnectorConfig.SNOWPIPE_STREAMING_CLIENT_PROVIDER_OVERRIDE_MAP);
@@ -844,13 +847,13 @@ public class ConnectorConfigTest {
     config.put(
         SnowflakeSinkConnectorConfig.SNOWPIPE_STREAMING_CLIENT_PROVIDER_OVERRIDE_MAP,
         "a:b,c:d,e:100,f:true");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testInvalidEmptyConfig() {
     Map<String, String> config = new HashMap<>();
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SNOWFLAKE_DATABASE)
         .hasMessageContaining(SNOWFLAKE_SCHEMA)
@@ -892,14 +895,14 @@ public class ConnectorConfigTest {
     config.put(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_ID, "client_id");
     config.put(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_SECRET, "client_secret");
     config.put(SnowflakeSinkConnectorConfig.OAUTH_REFRESH_TOKEN, "refresh_token");
-    Utils.validateConfig(config);
+    connectorConfigValidator.validateConfig(config);
   }
 
   @Test
   public void testInvalidAuthenticator() {
     Map<String, String> config = getConfig();
     config.put(SnowflakeSinkConnectorConfig.AUTHENTICATOR_TYPE, "invalid_authenticator");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.AUTHENTICATOR_TYPE);
   }
@@ -910,7 +913,7 @@ public class ConnectorConfigTest {
     config.put(SnowflakeSinkConnectorConfig.AUTHENTICATOR_TYPE, Utils.OAUTH);
     config.put(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_SECRET, "client_secret");
     config.put(SnowflakeSinkConnectorConfig.OAUTH_REFRESH_TOKEN, "refresh_token");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_ID);
   }
@@ -921,7 +924,7 @@ public class ConnectorConfigTest {
     config.put(SnowflakeSinkConnectorConfig.AUTHENTICATOR_TYPE, Utils.OAUTH);
     config.put(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_ID, "client_id");
     config.put(SnowflakeSinkConnectorConfig.OAUTH_REFRESH_TOKEN, "refresh_token");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_SECRET);
   }
@@ -932,7 +935,7 @@ public class ConnectorConfigTest {
     config.put(SnowflakeSinkConnectorConfig.AUTHENTICATOR_TYPE, Utils.OAUTH);
     config.put(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_ID, "client_id");
     config.put(SnowflakeSinkConnectorConfig.OAUTH_CLIENT_SECRET, "client_secret");
-    assertThatThrownBy(() -> Utils.validateConfig(config))
+    assertThatThrownBy(() -> connectorConfigValidator.validateConfig(config))
         .isInstanceOf(SnowflakeKafkaConnectorException.class)
         .hasMessageContaining(SnowflakeSinkConnectorConfig.OAUTH_REFRESH_TOKEN);
   }
@@ -944,7 +947,7 @@ public class ConnectorConfigTest {
     }
 
     try {
-      Utils.validateConfig(config);
+      connectorConfigValidator.validateConfig(config);
     } catch (SnowflakeKafkaConnectorException exception) {
       for (String configParam : paramsToRemove) {
         assert exception.getMessage().contains(configParam);
